requirements.txt
Django==5.0.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.32.3
fpdf2==2.7.8
psycopg2-binary
gunicorn==22.0.0
whitenoise==6.7.0
